pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = ""
authors = ["Sherwin-14 <varghesesheriwn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
tensorflow = "^2.17.0"
pydot = "^3.0.1"
pillow = "^10.4.0"
tqdm = "^4.66.4"
scikit-learn = "^1.5.1"
streamlit = "^1.36.0"
annoy = "^1.17.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
